refactor(userspace): extract HA/session-sync wire types to protocol_ha.go
Tenth code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the chassis-cluster HA state-update request, HA group
status, and the incremental session-sync/export/delta wire types into a
cohesive sibling protocol_ha.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change):
HAStateUpdateRequest, HAGroupStatus, SessionDeltaDrainRequest,
SessionExportRequest, SessionSyncRequest, SessionDeltaInfo.

The time import migrates with SessionDeltaInfo's time.Time field.
Behavior byte-identical; json tags preserved for the Rust userspace-dp
session-sync wire. Build passes; gofmt clean.
